azure-uploader/requirements.txt
azure-storage-blob>=12.19.0
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.66.0
pytest>=7.0.0
pytest-mock>=3.10.0
